Monobehaviour Event Listener Base
This is the base class for your daughters.

It has 3 methods, precisely to register the methods that will be called, remove them and finally execute the event, thus calling all registered methods.

So, here we see, I can use the same code for the same GameObjects and they will be executed only by themselves, without affecting others in the scene.